feat: refactor asset table to seperate components (mobile/desktop)
