+++ ui/app/assetTable/index.ts
@@ -1,1 +1,3 @@
+export * from "./desktopAssetTable"
+export * from "./mobileAssetTable"
 export * from "./assetTable"
